Create Watersheds.R
This code extracts the shapefiles for 102 of the 116 watersheds used in this analysis. This code is very temperamental such that it may change for each computer it is run on. It is chunked out as it is because of size limit errors I was getting when running the code in larger batches. Some watersheds need to be run on their own while others can be run in groups.